@@ -14,7 +14,7 @@ namespace DrawTest
 		/// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
 		protected override void Dispose(bool disposing)
 		{
-			if (disposing && (components != null))
+			if (disposing && (components is not null))
 			{
 				components.Dispose();
 			}
