@@ -13,7 +13,7 @@
 		/// <param name="disposing">true si les ressources managées doivent être supprimées ; sinon, false.</param>
 		protected override void Dispose(bool disposing)
 		{
-			if (disposing && (components != null))
+			if (disposing && (components is not null))
 			{
 				components.Dispose();
 			}
